server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legal-qa-server"
version = "0.1.0"
description = "Legal QA Server with RAG Pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.14",
    "langchain-core>=0.2.39",
    "langchain>=0.2.14",
    "groq>=0.13.0",
    "cohere>=5.5.0",
    "pinecone-client>=3.2.0",
    "httpx>=0.28.0",
    "tiktoken>=0.7.0",
    "tqdm>=4.66.0",
    "requests>=2.32.0",
    "numpy>=1.26.0",
    "pypdf>=4.2.0",
    "llama-index>=0.10.0",
    "llama-index-embeddings-cohere>=0.1.0",
    "llama-index-vector-stores-pinecone>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
[tool.hatch.build.targets.wheel]
packages = ["src"]